Add files via upload
Abstract: Americans’ confidence in the medical and scientific communities have dwindled over the last several decades. With data from the 2018 and 2021 General Social Survey, I examine the effect of (lack of) confidence in these communities on individuals’ health outcomes before and during the COVID-19 pandemic using ordered logistic regression and heterogeneous choice models. The results show that counterintuitively, COVID-19 did not have a significant impact on the confidence in these communities and none of the research hypotheses were accepted, which indicates the need to assess the larger societal landscape from the past 50 years.